+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file. The format 
 
 ## Unreleased
 
+- Added a documentation website built with [Documenter.jl](https://github.com/JuliaDocs/Documenter.jl) and [Literate.jl](https://github.com/fredrikekre/Literate.jl).
+
 ## 3.2.0
 
 - The `RFAM_DIR` and `RFAM_VERSION` preferences can now be read from environment variables `ENV["JULIA_RFAM_DIR"]` and `ENV["JULIA_RFAM_VERSION"]`, if they exist. If the `LocalPreferences.toml` file is present specifying these values, it takes precedence and the environment variables will be ignored.